# api/resource/swagger/FeatureManager-swagger.json
{
    "swagger": "2.0",
    "paths": {
...
        "/features/{key}": {
            "delete": {
                "parameters": [
                    {
                        "name": "key",
                        "type": "string",
                        "in": "path",
                        "required": true,
                        "description": null
                    }
                ],
                "tags": [
                    "Feature"
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    }
                }
            },
            "get": {
                "parameters": [
                    {
                        "name": "key",
                        "type": "string",
                        "in": "path",
                        "required": true,
                        "description": null
                    }
                ],
                "tags": [
                    "Feature"
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "$ref": "#/definitions/FeatureResponseDto"
                        }
                    }
                }
            },
            "post": {
                "parameters": [
                    {
                        "name": "key",
                        "type": "string",
                        "in": "path",
                        "required": true,
                        "description": null
                    },
                    {
                        "name": "body",
                        "type": "object",
                        "in": "body",
                        "required": true,
                        "description": null,
                        "schema": {
                            "$ref": "#/definitions/FeatureRequestDto"
                        }
                    }
                ],
                "tags": [
                    "Feature"
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "$ref": "#/definitions/FeatureResponseDto"
                        }
                    }
                }
            },
            "put": {
                "parameters": [
                    {
                        "name": "Authorization",
                        "description": "Bearer <token>",
                        "in": "header",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "key",
                        "type": "string",
                        "in": "path",
                        "required": true,
                        "description": null
                    },
                    {
                        "name": "body",
                        "type": "object",
                        "in": "body",
                        "required": true,
                        "description": null,
                        "schema": {
                            "$ref": "#/definitions/FeatureRequestDto"
                        }
                    }
                ],
                "tags": [
                    "Feature"
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "$ref": "#/definitions/FeatureResponseDto"
                        }
                    }
                }
            }
        },
...
        "/feature-datas/{featureKey}/{sampleKey}": {
            "delete": {
                "parameters": [
                    {
                        "name": "featureKey",
                        "type": "string",
                        "in": "path",
                        "required": true,
                        "description": null
                    },
                    {
                        "name": "sampleKey",
                        "type": "string",
                        "in": "path",
                        "required": true,
                        "description": null
                    }
                ],
                "tags": [
                    "FeatureData"
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    }
                }
            },
            "get": {
                "parameters": [
                    {
                        "name": "featureKey",
                        "type": "string",
                        "in": "path",
                        "required": true,
                        "description": null
                    },
                    {
                        "name": "sampleKey",
                        "type": "string",
                        "in": "path",
                        "required": true,
                        "description": null
                    }
                ],
                "tags": [
                    "FeatureData"
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "$ref": "#/definitions/FeatureDataResponseDto"
                        }
                    }
                }
            }
        },
        "/feature-datas/batch/{featureKey}": {
            "get": {
                "parameters": [
                    {
                        "name": "featureKey",
                        "type": "string",
                        "in": "path",
                        "required": true,
                        "description": null
                    }
                ],
                "tags": [
                    "FeatureData"
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "object",
                                "$ref": "#/definitions/FeatureDataResponseDto"
                            }
                        }
                    }
                }
            }
        },
...
        "/samples/{key}": {
            "delete": {
                "parameters": [
                    {
                        "name": "key",
                        "type": "string",
                        "in": "path",
                        "required": true,
                        "description": null
                    }
                ],
                "tags": [
                    "Sample"
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    }
                }
            },
            "get": {
                "parameters": [
                    {
                        "name": "key",
                        "type": "string",
                        "in": "path",
                        "required": true,
                        "description": null
                    }
                ],
                "tags": [
                    "Sample"
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "$ref": "#/definitions/SampleResponseDto"
                        }
                    }
                }
            },
            "post": {
                "parameters": [
                    {
                        "name": "key",
                        "type": "string",
                        "in": "path",
                        "required": true,
                        "description": null
                    },
                    {
                        "name": "body",
                        "type": "object",
                        "in": "body",
                        "required": true,
                        "description": null,
                        "schema": {
                            "$ref": "#/definitions/SampleRequestDto"
                        }
                    }
                ],
                "tags": [
                    "Sample"
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "$ref": "#/definitions/SampleResponseDto"
                        }
                    }
                }
            },
            "put": {
                "parameters": [
                    {
                        "name": "key",
                        "type": "string",
                        "in": "path",
                        "required": true,
                        "description": null
                    },
                    {
                        "name": "body",
                        "type": "object",
                        "in": "body",
                        "required": true,
                        "description": null,
                        "schema": {
                            "$ref": "#/definitions/SampleRequestDto"
                        }
                    }
                ],
                "tags": [
                    "Sample"
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "$ref": "#/definitions/SampleResponseDto"
                        }
                    }
                }
            }
        },
        "/samples/{key}/{value}": {
            "patch": {
                "parameters": [
                    {
                        "name": "key",
                        "type": "string",
                        "in": "path",
                        "required": true,
                        "description": null
                    },
                    {
                        "name": "value",
                        "type": "integer",
                        "in": "path",
                        "required": true,
                        "description": null
                    },
                    {
                        "name": "body",
                        "type": "object",
                        "in": "body",
                        "required": true,
                        "description": null,
                        "schema": {
                            "$ref": "#/definitions/SampleRequestDto"
                        }
                    }
                ],
                "tags": [
                    "Sample"
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "$ref": "#/definitions/SampleResponseDto"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "FeatureResponseDto": {
            "type": "object",
            "properties": {
                "iterationCount": {
                    "type": "object",
                    "example": null
                },
                "key": {
                    "type": "object",
                    "example": null
                },
                "label": {
                    "type": "object",
                    "example": null
                }
            },
            "required": [
                "iterationCount",
                "key",
                "label"
            ]
        },
        "FeatureRequestDto": {
            "type": "object",
            "properties": {
                "label": {
                    "type": "object",
                    "example": null
                }
            },
            "required": [
                "label"
            ]
        },
        "FeatureDataResponseDto": {
            "type": "object",
            "properties": {
                "featureKey": {
                    "type": "object",
                    "example": null
                },
                "iterationCount": {
                    "type": "object",
                    "example": null
                },
                "sampleKey": {
                    "type": "object",
                    "example": null
                },
                "value": {
                    "type": "object",
                    "example": null
                }
            },
            "required": [
                "featureKey",
                "iterationCount",
                "sampleKey",
                "value"
            ]
        },
...
        "SampleResponseDto": {
            "type": "object",
            "properties": {
                "featureDataList": {
                    "type": "array",
                    "items": {
                        "type": "object",
                        "$ref": "#/definitions/FeatureDataResponseDto"
                    }
                },
                "iterationCount": {
                    "type": "object",
                    "example": null
                },
                "key": {
                    "type": "object",
                    "example": null
                },
                "label": {
                    "type": "object",
                    "example": null
                },
                "value": {
                    "type": "object",
                    "example": null
                }
            },
            "required": [
                "featureDataList",
                "iterationCount",
                "key",
                "label",
                "value"
            ]
        },
        "SampleRequestDto": {
            "type": "object",
            "properties": {
                "featureDataList": {
                    "type": "array",
                    "items": {
                        "type": "object",
                        "$ref": "#/definitions/FeatureDataRequestDto"
                    }
                },
                "label": {
                    "type": "object",
                    "example": null
                }
            },
            "required": [
                "featureDataList",
                "label"
            ]
        },
        "FeatureDataRequestDto": {
            "type": "object",
            "properties": {
                "featureKey": {
                    "type": "object",
                    "example": null
                },
                "sampleKey": {
                    "type": "object",
                    "example": null
                }
            },
            "required": [
                "featureKey",
                "sampleKey"
            ]
        },
...
    },
...
    "host": "localhost:5030",
    "schemes": [
        "http"
    ],
    "basePath": "/feature-dataset-manager-api",
    "info": {
        "title": "FeatureManager",
        "description": "This is an api that manage samples and its features. The main goal is to retrieve the best sample match given a set of features",
        "version": "0.0.1",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "Samuel Jansen",
            "email": "samuel.jansenn@gmail.com"
        },
        "license": {
            "name": "Apache 2.0 / MIT License",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        }
    }
}